COM-3408 - create AttendanceSheets composable
